pyproject.toml
[project]
name = "capstone-optimal-bidding"
version = "0.1.0"
description = "MTHE 493 Capstone - In Partnership With Enverus"
authors = [
    {name = "Thomas Boyle",email = "tsboyle10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
